Replace the mpool code with a new, simpler allocator.  SANDmark 5.4; and future work exists in inlining the eminently specializable (and simple!) alloc/free functions.
Extra advantage: no more dependence on fast bit-scan!

Drawback is that it's more prone to fragmentation; on the other hand, it
doesn't lose words to power-of-two roundup; on the third hand, it has
to prefix everything with a length; on the fourth hand, it can use any
old hunk of memory instead of needing one specially mapped, so large
allocations are handed off to malloc and then freed normally.

Future work may also exist in increasing fragmentation resistance; I
have reason to believe that this can be kept out of the fast path.