Add .c.o rule for what GNU make had been (nonstandardly) giving me.
POSIX specifies no "-o $@", which means i386/co.c will yield co.o instead
of i386/co.o, which is what the target is supposed to be producing.  Not
the most useful thing in the world, but it is the standard, and other
makes (e.g., BSD) implement it as written.